removed the passwords hash from the list of users
Removed the password from list of users

+++ routes/dvd.js
@@ -1,13 +1,12 @@
 /*
 created by Toby Cantello
 created on 10/4/23
-updated on 10/25/23
+updated on 11/7/23
 */
 
 const express = require ('express');
 const {check, validationResult} = require("express-validator");
 const authUser = require('../midWare/authLogin.js');
-
 const router = express.Router();
 router.use(express.json());
 router.use(express.urlencoded({ extended: true }));
@@ -18,8 +17,15 @@ const DVD = require('../models/DVD.js');
 router.get("/", authUser, async (req, res, next) => {
     try 
         {
-            const dvds = await DVD.findAll();
-            res.send(dvds);
+            if (loggedIn == 1)
+                {
+                    const dvds = await DVD.findAll();
+                    res.send(dvds);
+                }
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }
         }
     catch (error)
         {
@@ -28,62 +34,87 @@ router.get("/", authUser, async (req, res, next) => {
 });
 
 // GET route to get the dvds in the database with the matching ID
-router.get("/:id", async (req, res, next) => {
+router.get("/:id", authUser, async (req, res, next) => {
     try
         {
-            const dvds = await DVD.findOne({where: {id: req.params.id}});
-            res.send(dvds);
+            if (loggedIn == 1)
+                {
+                    const dvds = await DVD.findOne({where: {id: req.params.id}});
+                    res.send(dvds);
+                }
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }
         }
     catch (error)
-
         {
             next((error));
         }
 })
 
 // GET route to get the dvds in the database with the matching ID
-router.get("/titles/:title", async (req, res, next) => {
+router.get("/titles/:title", authUser, async (req, res, next) => {
     try
         {
-            const dvds = await DVD.findAll({where: {title: req.params.title}});
-            res.send(dvds);
+            if (loggedIn == 1)
+                {
+                    const dvds = await DVD.findAll({where: {title: req.params.title}});
+                    res.send(dvds);
+                }
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }
         }
     catch (error)
-
         {
             next((error));
         }
 })
 
 // GET route to get the dvds in the database with the matching year
-router.get("/years/:year", async (req, res, next) => {
+router.get("/years/:year", authUser, async (req, res, next) => {
     try
         {
-            const dvds = await DVD.findAll({where: {year: req.params.year}});
-            res.send(dvds);
+            if (loggedIn == 1)
+                {
+                    const dvds = await DVD.findAll({where: {year: req.params.year}});
+                    res.send(dvds);
+                }
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }
         }
     catch (error)
-
         {
             next((error));
         }
 })
 
 // POST route to create a dvd in the database
-router.post('/', [check("title").not().isEmpty().trim(), check("director").not().isEmpty().trim(), check("year").not().isEmpty().trim()], async (req, res, next) => {
+router.post('/', authUser, [check("title").not().isEmpty().trim(), check("director").not().isEmpty().trim(), check("year").not().isEmpty().trim()], async (req, res, next) => {
     try 
         {
-            const errors = validationResult(req);
+            if (loggedIn == 1)
+                {
+                    const errors = validationResult(req);
 
-            if(!errors.isEmpty())
-                {
-                    res.json({ error: errors.array() })
+                    if(!errors.isEmpty())
+                        {
+                            res.json({ error: errors.array() })
+                        }
+                    else 
+                        {
+                            const dvds = await DVD.create(req.body);
+                            res.send(dvds);
+                        }
                 }
-            else 
-                {
-                    const dvds = await DVD.create(req.body);
-                    res.send(dvds);
-                } 
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }  
         }
     catch (error)
         {
@@ -92,11 +123,18 @@ router.post('/', [check("title").not().isEmpty().trim(), check("director").not()
 });
 
 // PUT route to update a dvd in the database
-router.put('/:id', async (req, res, next) => {
+router.put('/:id', authUser, async (req, res, next) => {
     try
         {
-            const dvd = await DVD.update(req.body, {where: {id: req.params.id}});
-            res.sendStatus(200);
+            if (loggedIn == 1)
+                {
+                    const dvd = await DVD.update(req.body, {where: {id: req.params.id}});
+                    res.sendStatus(200);
+                }
+            else
+            {
+                res.send({message: "Please login to see this information"})
+            }
         }
     catch (error)
         {
@@ -105,22 +143,30 @@ router.put('/:id', async (req, res, next) => {
 });
 
 // DELETE route to delete dvd with matching ID
-router.delete('/:id', async (req, res, next) => {
+router.delete('/:id', authUser, async (req, res, next) => {
     try 
         {
-            const dvd = await DVD.destroy({where: {id: req.params.id}});
-            if (dvd === 0) 
+            if (loggedIn == 1)
                 {
-                    throw new Error("No DVD to deleted");
+                    const dvd = await DVD.destroy({where: {id: req.params.id}});
+                    if (dvd === 0) 
+                        {
+                            throw new Error("No DVD to deleted");
+                        }
+                    else
+                        {
+                            res.sendStatus(200);
+                        }
                 }
             else
                 {
-                    res.sendStatus(200);
-                }   
+                    res.send({message: "Please login to see this information"})
+                } 
         } 
-    catch (error) {
-        next(error);
-    }
+    catch (error) 
+        {
+            next(error);
+        }
   });
 
 module.exports = router;
